Improved upload center experience, code cleanup
